remove unnecessary waiting when pxeboot fails
In the instances we have not found the ETI stub message or have not been able to successfully ping the bf by a certain point, waiting longer does not seem to result in success. It would seem the issue is not timing out, but rather something that has already gone wrong with the boot / race condition with DHCP.

In the case of looking for the ETI stub message when pxe-booting, there will be early output we can detect prior to the ETF stub message which shows the pxe-boot is progressing ("Station IP address *"). In the event we don't see this output, the pxeboot has already failed, so we can bail out earlier and retry to reduce the time spent waiting.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>