dpu-tools: rework run() method
This is copied over from common_bf.run(). The difference is that it better
handles typing.

Later, we should not copy code around but re-use code (e.g. by importing
the common_bf module). But for now, it is still copied because to import
the code we also need to ensure that we include the common_bf module.