common_ipu: Add shareable common functionality between multiple modules
- Modified and moved run into common_ipu:
This run command is able to both capture output and stream the output to
the screen since long commands like with flashing the ipu with dd, feed
back is essential to help the user not interrupt its task.
After a lot of trial and error this function needs threads since reading
from both stdout and stderr is inherently a blocking operation which
would halt the output until the end of the command. Likewise, passing
the shell=True argument to Popen worked better than parsing the command
with shlex since many commands run in fwutils requires piping commands
and ssh which with a shell, it's handled easily.

- Added a Version list constant to track the supported version which is
useful for the flags choices

- Moved the Result structure to common_ipu to be shared between modules

- Added get_current_version that ssh into the IMC and runs cat
/etc/issue.net and parses the version, returning it in a Result

- Added a minicom library for dpu-tools

Since minicom functionality seems to be useful to multiples components
in dpu-tools for mev 1.2, keeping it in the minicom.py helps share some
code.

- Added a download files function

It would be useful to be able to download a file in a temp location that
can be used to then flash onto the ipu, a nice way was using the
requests library, since we can have easier error handling, chunck large
downloads and etc.

- Added an extracting compressed tar function

Since we have a server already to serve the ISOs, hosting the IPU
firmware images on that server is a straight forward way to have a
repository. In that server there a good way to transferring the images
quickly is to have it compressed; so this function can take that
downloaded compressed tar and extract its contents to be used in the
flashing steps

- Added a function to find the ipu firmware images needed for each
flashing step

The ipu firmware files follow a specific structure, so this function
allows us to automate the process of getting the necessary bin files
depending on the version needed. Essentially we can use this file in
get_images to allow reset to infer what the files it needs depending on
the version given by get_version in fwutils

- Added a function to setup logging based on verbosity

Keeps the supporting modules in sync with fwutils to only display useful
output depending on the users needs.

- Added get_version where it ssh into the IMC and runs cat
  /etc/issue.net to parse its contents and return its result.

- Added a minicom version of get_version where it runs cat
  /etc/issue.net and parses it to find the current version that can be
run
  before IMC connectivity is enabled.