minicom.py: Add a minicom module
This minicom module is useful for getting IMC connectivity in Ipu MeV
1.2
- added minicom_cmd to return the correct minicom command based on if
its an imc or the acc.
This can later be used for parts of the dpu-tools like in the console
command

- added pexpect_child_wait to be used in command functions that require
waiting for a specific pattern like enabling connectivity in the IMC,
this function is borrowed by the pxeboot module.

- added configure_minicom to fix the problem where the minicom can't
write any input because of Hardware Flow Control