fwutils.py: Change Simplify Behavior of IPU Runtime Access step
We should avoid minicom as much as possible as it can be hard to predict
all the edge cases needed to successfully perform what we have in mind.
In the IPU Runtime Access step it would mean we should first check if it
can connect to it properly, then if not, try the more "manual" method.