common_ipu.py: Add check_connectivity function
We should have a simple way to check connectivity to a dpu by pinging
its address. This function could be used in the IPU Runtime Access step
to not to attempt running minicom if it already can stablish a
connection with the dpu therefore avoiding cases where the runtime
access has already been granted which minicom may not know how to handle
properly.